requirements.txt
altgraph==0.17
asgiref==3.3.1
async-timeout==3.0.1
atlastk==0.12.9
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-filter==2.4.0
django-heroku==0.3.1
django-jsonfield==1.4.1
django-model-utils==4.1.1
docutils==0.17
Eel==0.12.4
fields==5.0.0
future==0.18.2
gevent==21.1.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
legacy==0.1.6
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.3
pefile==2019.4.18
Pillow==8.1.0
portalocker==2.3.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
s3transfer==0.3.7
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
txaio==21.2.1
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.2.0
python-decouple==3.4
django-googledrive-storage==1.5.2
cloudinary==1.26.0
django-cloudinary-storage==0.3.0